pyworker/requirements.txt
annoy==1.17.3
asyncio==3.4.3
facenet-pytorch==2.5.3
grpcio==1.70.0
grpcio-tools==1.70.0
minio==7.2.15
moviepy==1.0.3
numpy==2.2.3
opencv-python==4.11.0.86
prometheus-client==0.21.1
Pillow==10.4.0
PyExifTool==0.5.6
rapidocr-onnxruntime==1.4.4
rawpy==0.24.0
requests==2.32.3
schedule==1.2.2
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
Wand==0.6.13
